% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athletes-package.R
\docType{package}
\name{athletes-package}
\alias{athletes}
\alias{athletes-package}
\title{athletes: IND Athletes Analysis}
\description{
Tools and analyses from the IND athletes dataset.
}
\author{
\strong{Maintainer}: Matías Castillo-Aguilar \email{matcasti@umag.cl} (\href{https://orcid.org/0000-0001-7291-247X}{ORCID})

}
\keyword{internal}
